test: Refactored unittest configuration
