+++ test/conftest.py
@@ -1,0 +1,23 @@
+import pytest
+import os
+import requests
+from io import BytesIO
+
+# Disable GPU for testing
+os.environ["CUDA_VISIBLE_DEVICES"] = "-1"
+
+
+@pytest.fixture(scope="session")
+def mock_vocab():
+    return ('3K}7eé;5àÎYho]QwV6qU~W"XnbBvcADfËmy.9ÔpÛ*{CôïE%M4#ÈR:g@T$x?0î£|za1ù8,OG€P-kçHëÀÂ2É/ûIJ\'j'
+            '(LNÙFut[)èZs+&°Sd=Ï!<â_Ç>rêi`l')
+
+
+@pytest.fixture(scope="session")
+def mock_pdf(tmpdir_factory):
+    url = 'https://arxiv.org/pdf/1911.08947.pdf'
+    file = BytesIO(requests.get(url).content)
+    fn = tmpdir_factory.mktemp("data").join("mock_pdf_file.pdf")
+    with open(fn, 'wb') as f:
+        f.write(file.getbuffer())
+    return str(fn)
