+++ test/test_models.py
@@ -1,15 +1,11 @@
 import pytest
-import os
 import numpy as np
 import tensorflow as tf
 
-os.environ["CUDA_VISIBLE_DEVICES"] = "-1"
-
+from doctr import models
 from doctr.documents import read_pdf, Document
-from test_documents_reader import mock_pdf
 from test_models_detection import test_detectionpredictor
 from test_models_recognition import test_recognitionpredictor
-from doctr import models
 
 
 def test_extract_crops(mock_pdf):  # noqa: F811
