+++ test/test_models_recognition.py
@@ -1,21 +1,11 @@
 import pytest
-import os
 import numpy as np
 import math
 import tensorflow as tf
 
-os.environ["CUDA_VISIBLE_DEVICES"] = "-1"
-
 from doctr.models import recognition
 from doctr.documents import read_pdf
 from doctr.models import extract_crops
-from test_documents_reader import mock_pdf
-
-
-@pytest.fixture(scope="session")
-def mock_vocab():
-    return ('3K}7eé;5àÎYho]QwV6qU~W"XnbBvcADfËmy.9ÔpÛ*{CôïE%M4#ÈR:g@T$x?0î£|za1ù8,OG€P-kçHëÀÂ2É/ûIJ\'j'
-            '(LNÙFut[)èZs+&°Sd=Ï!<â_Ç>rêi`l')
 
 
 def test_recopreprocessor(mock_pdf):  # noqa: F811
